requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-groq
langchain-openai
pyyaml==6.0.1